pyproject.toml
[tool.poetry]
name = "File Upload Service"
version = "0.1.0"
description = ""
authors = ["anandafiqri <anandafiqri1403@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"
imagekitio = "^4.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
